requirements.txt
streamlit>=1.25
requests>=2.31
Pillow>=9.5
diskcache>=5.6